requirements.txt
# Hyperfocus Backend Dependencies
# Phase 1 - Minimal viable backend

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# HTTP client (for vLLM proxy)
httpx==0.27.2

# Database
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0

# WebSocket support (for terminal)
websockets==13.0.1

# Phase 2 (uncomment when ready)
# asyncpg==0.29.0  # Postgres async driver
# pgvector==0.3.2  # Vector search
# sse-starlette==2.1.3  # Server-sent events

# Document generation
python-docx==1.1.0
markdown==3.5.2
weasyprint==62.0
pypdf==4.0.2
